Class to implement 'plans'.  Takes same base directory and type subdirs as Network, plus tag and tempdir.  These are necessary because this class checks out projects in order to call the getNetTypes function and return a list of projects of a user-specified type.
A 'plan' is defined by a specification file with two columns.  The first is the name of the project, and the second is the project type: 'plan' or 'project'.  This class reads the spec file, builds a list of projects to be included, and asks the project what type of networks it belongs to.

listOfProjects allows user to get a list of projects for a specified network type.

Signed-off-by: Drew <drew.cooper@sfcta.org>